refactor: simplifies reducer
